# public/api-docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "\n\nHeader -> api_key :  \n\n",
    "version": "1.0.0",
    "title": "E-commerce"
  },
  "host": "127.0.0.1:8000",
  "basePath": "/api",
  "tags": [
    {
      "name": "User",
      "description": "All User Operations"
    },
        {
      "name": "Home Page Data",
      "description": "All Home Page Data"
    }
  
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "User Login",
        "operationId": "register",
        "produces": [ 
          "application/json"
        ],
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
             {
            "name": "name",
            "in": "formData",
            "description": "Enter Name",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "Enter Email",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "password",
            "required": true,
            "type": "string"
          }  
         
        ],
        "responses": {
          "200": {
            "description": "Data"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "User Login",
        "operationId": "login",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          
          {
            "name": "email",
            "in": "formData",
            "description": "Enter Email",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "password",
            "required": true,
            "type": "string"
          }  
         
        ],
        "responses": {
          "200": {
            "description": "Data"
          }
        }
      }
    },
     "/user": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Show User Data",
        "operationId": "showUser",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          
          {
            "name": "Authorization",
            "in": "header",
            "description": "Enter Bearer Token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Data"
          }
        }
      }
    },
    "/updateUser": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Show User Data",
        "operationId": "updateUser",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          
          {
            "name": "Authorization",
            "in": "header",
            "description": "Enter Bearer Token",
            "required": true,
            "type": "string"
          }  ,
             {
            "name": "name",
            "in": "formData",
            "description": "Enter Name",
            "required": true,
            "type": "string"
          }  
         
        ],
        "responses": {
          "200": {
            "description": "Data"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Show User Data",
        "operationId": "updateUser",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          
          {
            "name": "Authorization",
            "in": "header",
            "description": "Enter Bearer Token",
            "required": true,
            "type": "string"
          }  
         
        ],
        "responses": {
          "200": {
            "description": "Data"
          }
        }
      }
    } ,
     "/getHomeData": {
      "get": {
        "tags": [
          "Home Page Data"
        ],
        "summary": "Home Page Data",
        "operationId": "getHomeData",  
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          
      
         
        ],
        "responses": {
          "200": {
            "description": "Data"
          }
        }
      }
    } ,
     "/getCategoryData/{slug}": {
      "get": {
        "tags": [
          "Category page Data"
        ],
        "summary": "Category page Data",
        "operationId": "getCategoryData",  
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
           {
            "name": "slug",
            "in": "path",
            "description": "Enter slug",
            "required": true,
            "type": "string"
          }  
        
        ],
        "responses": {
          "200": {
            "description": "Data"
          }
        }
      }
    } ,
     "/getHeaderCategoriesData": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Home Page Data",
        "operationId": "getHeaderCategoriesData",  
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          
      
         
        ],
        "responses": {
          "200": {
            "description": "Data"
          }
        }
      }
    }                
    
   
  },
  "securityDefinitions": {
    "APIKeyHeader": {
      "type": "apiKey",
      "in": "header",
      "name": "api-key"
    }
  },
  "definitions": {
    "user": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "number"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    }
  },
  "schemes": [
    "http",
    "https"
  ]
}